% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildLLR.R
\name{buildLLR.gauss}
\alias{buildLLR.gauss}
\title{Build LLR function using Gaussian densities.}
\usage{
buildLLR.gauss(posScores, negScores, spline = TRUE)
}
\arguments{
\item{posScores}{fitness scores in the positive reference set}

\item{negScores}{fitness scores in the negative reference set}

\item{spline}{logical; whether or not to apply spline monotonization. TRUE by default.}
}
\value{
a list containing the LLR function in log10 scale and the positive and negative density functions
}
\description{
Build LLR function using Gaussian densities.
}
\examples{

posScores <- rnorm(50,0.2,0.3)
negScores <- rnorm(30,0.8,0.2)
llr <- buildLLR.gauss(posScores,negScores)
drawDensityLLR(c(posScores,negScores),llr$llr,llr$posDens,llr$negDens,posScores,negScores)

}
