% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildLLR.R
\name{buildLLR.kernel}
\alias{buildLLR.kernel}
\title{Build LLR function using kernel density estimation}
\usage{
buildLLR.kernel(posScores, negScores, bw = 0.1, kernel = "gaussian")
}
\arguments{
\item{posScores}{fitness scores in the positive reference set}

\item{negScores}{fitness scores in the negative reference set}

\item{bw}{the bandwith to use for kernel density estimation (see kdensity package for more info).
This can either be a numerical value or the name of the algorithm used to automatically choose one.}

\item{kernel}{the type of kernel to use (see kdensity package for more info)}
}
\value{
a list containing the LLR function in log10 scale and the positive and negative density functions
}
\description{
Build LLR function using kernel density estimation
}
\examples{
posScores <- rnorm(50,0.2,0.3)
negScores <- rnorm(30,0.8,0.2)
llr <- buildLLR.kernel(posScores,negScores,bw=0.1,kernel="gaussian")
drawDensityLLR(c(posScores,negScores),llr$llr,llr$posDens,llr$negDens,posScores,negScores)

}
