% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildLLR.R
\name{drawDensityLLR}
\alias{drawDensityLLR}
\title{Draw a plot that summarizes the densities and LLR calculated by other functions}
\usage{
drawDensityLLR(scores, llrFun, posDens, negDens, posScores, negScores)
}
\arguments{
\item{scores}{the numerical scores in the Mave map}

\item{llrFun}{the LLR function}

\item{posDens}{the density function of the positive reference set}

\item{negDens}{the density function of the negative reference set}

\item{posScores}{the numerical scores in the postive reference set}

\item{negScores}{the numerical scores in the negative reference set}
}
\value{
nothing
}
\description{
Draw a plot that summarizes the densities and LLR calculated by other functions
}
\examples{
posScores <- rnorm(50,0.2,0.3)
negScores <- rnorm(30,0.8,0.2)
llr <- buildLLR.gauss(posScores,negScores)
drawDensityLLR(c(posScores,negScores),llr$llr,llr$posDens,llr$negDens,posScores,negScores)

}
