% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildLLR.R
\name{drawDensityLLR_fixedRange}
\alias{drawDensityLLR_fixedRange}
\title{Like drawDensityLLR, but with fixed y range for better comparison across different LLRs}
\usage{
drawDensityLLR_fixedRange(
  scores,
  llrFun,
  posDens,
  negDens,
  posScores,
  negScores,
  prior = 0.1
)
}
\arguments{
\item{scores}{the numerical scores in the Mave map}

\item{llrFun}{the LLR function}

\item{posDens}{the density function of the positive reference set}

\item{negDens}{the density function of the negative reference set}

\item{posScores}{the numerical scores in the postive reference set}

\item{negScores}{the numerical scores in the negative reference set}
}
\value{
nothing
}
\description{
Like drawDensityLLR, but with fixed y range for better comparison across different LLRs
}
\examples{
posScores <- rnorm(50,0.2,0.3)
negScores <- rnorm(30,0.8,0.2)
llr <- buildLLR.gauss(posScores,negScores)
drawDensityLLR_fixedRange(c(posScores,negScores),llr$llr,llr$posDens,llr$negDens,posScores,negScores)

}
