% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildLLR.R
\name{findLLRcrossings}
\alias{findLLRcrossings}
\title{Find x-values at which the given LLR function crosses specified thresholds.}
\usage{
findLLRcrossings(llrFun, thresholds, xlim, nPoints = 1000)
}
\arguments{
\item{llrFun}{the LLR Function}

\item{thresholds}{thresholds for each ACMG category}

\item{xlim}{range of x values i.e. scores}

\item{nPoints}{an integer specifying number of points to sample across xlim}
}
\value{
a table with threshold labels, values, and x when crossing
}
\description{
Find x-values at which the given LLR function crosses specified thresholds.
}
\examples{
# Using a previously defined LLR function
posScores <- rnorm(50,0.2,0.3)
negScores <- rnorm(30,0.8,0.2)
llrList <- buildLLR.kernel(posScores, negScores, bw=0.1, kernel="gaussian")
llrFun <- llrList$llr

# Define thresholds of interest
llrTs <- llrThresholds(optiLLR(0.1))

# Find the crossings in the range of observed scores
x_range <- range(c(posScores, negScores))
findLLRcrossings(llrFun, llrTs, xlim = x_range)
}
